+++ e-learning/src/components/Button.module.scss
@@ -1,33 +1,55 @@
 .container {
-  color: white;
-  font-weight: 500;
+  font-weight: 400;
   display: inline-flex;
-  padding: 0.3em 0.75em;
-  padding-bottom: 0.4em;
-  border-radius: 8px;
+  padding: 0.35em 0.65em;
+  border-radius: 6px;
   text-decoration-width: 1px;
   cursor: pointer;
   gap: 0.25em;
+  transition: var(--fast);
   align-items: baseline;
 }
 
 .primary {
   background: var(--blue-90);
+  color: white;
   &:hover,
   &:focus-visible {
     background: var(--blue-80);
   }
 }
 .secondary {
+  background: var(--gray-20);
+  border: 1px solid var(--gray-20);
+  color: var(--blue-90);
+  &:hover,
+  &:focus-visible {
+    border: 1px solid var(--blue-80);
+  }
+}
+.tertiary {
+  background: var(--gray-20);
+  color: var(--blue-90);
 }
 
 .has-icon {
   padding-right: 0.5em;
 }
 
+.hide-label {
+  padding: 0.3em;
+  .label {
+    position: absolute;
+    font-size: 0;
+    pointer-events: none;
+  }
+}
+
 .icon {
-  width: 1em;
-  height: auto;
   position: relative;
-  top: 0.15em;
+  svg {
+    height: 1.35em;
+    width: auto;
+    display: block;
+  }
 }
