+++ e-learning/src/components/ButtonAdapter.js
@@ -1,0 +1,7 @@
+import React from 'react'
+import SharedButton from '../../../shared/components/Button'
+import * as styles from './Button.module.scss'
+
+export default function Button(props) {
+  return <SharedButton styles={styles} {...props} />
+}
